+++ src/query.rs
@@ -1,2 +1,3 @@
 pub mod album;
 pub mod artist;
+pub mod audiobook;